refactor: extract ChatWindow component
